test(editor): Move reusable frontend test setup into @n8n/vitest-config (no-changelog)
editor-ui's src/__tests__/setup.ts had grown to 496 lines, ~350 of which are
generic jsdom patching that every frontend package needs and six of them were
re-implementing in fragments. Extract that into a new
@n8n/vitest-config/frontend-setup export and have all eight frontend setup
files consume it, so a scaffolded module package gets the whole harness in one
import line.

Left in the packages deliberately, because each is genuinely local:
- editor-ui: fake-indexeddb/core-js polyfills, the reka-ui popover mock
  (reka-ui is an editor-ui-only dependency and vi.mock resolves its specifier
  relative to the calling file), the app-grid/app-modals containers, the i18n
  boot, and the matchMedia override.
- design-system: the N8nPlugin registration, the is-emoji-supported mock, and
  the defaultPrevented MouseEvent/PointerEvent patches.

Framework boot (pinia, i18n) stays out of the shared harness on purpose:
@n8n/i18n already devDepends on @n8n/vitest-config, so a dependency the other
way would be a turbo build cycle.

Two harness fixes fall out of the move:
- matchMedia now answers `false` by default instead of `true`. Answering true
  to every query silently opted components into whichever branch a media query
  guards; it disabled N8nEmptyState's icon animation via
  `prefers-reduced-motion: reduce`. editor-ui keeps `true` locally.
- passWithNoTests moves into the shared frontend config, so a sparse package
  no longer fails when CI's 2-way shard exceeds its test-file count.

Also add packages/@n8n/vitest-config/ to janitor's GLOBAL_TRIGGER_PREFIXES: a
per-package setup file is already a bail-to-full-run trigger, and the harness
body must keep triggering one now that it lives here, or a change to it would
scope to zero test files and report a false green.
