+++ packages/frontend/@n8n/chat/src/__tests__/setup.ts
@@ -1,21 +1,1 @@
-import '@testing-library/jest-dom/vitest';
-import { configure } from '@testing-library/vue';
-
-configure({ testIdAttribute: 'data-test-id' });
-
-class ResizeObserverMock extends EventTarget {
-	constructor() {
-		super();
-	}
-
-	observe = vi.fn();
-
-	disconnect = vi.fn();
-
-	unobserve = vi.fn();
-}
-
-beforeEach(() => {
-	vi.stubGlobal('ResizeObserver', ResizeObserverMock);
-});
-afterEach(() => vi.unstubAllGlobals());
+import '@n8n/vitest-config/frontend-setup';
