+++ packages/frontend/@n8n/frontend-module-sdk/src/__tests__/setup.ts
@@ -1,4 +1,1 @@
-// Test setup entry required by the shared frontend vitest config
-// (`@n8n/vitest-config/frontend` points `setupFiles` here). The registry suites
-// need no global setup, so this file intentionally stays empty.
-export {};
+import '@n8n/vitest-config/frontend-setup';
