+++ packages/frontend/@n8n/frontend-utils/src/__tests__/setup.ts
@@ -1,4 +1,1 @@
-import '@testing-library/jest-dom/vitest';
-import { configure } from '@testing-library/vue';
-
-configure({ testIdAttribute: 'data-test-id' });
+import '@n8n/vitest-config/frontend-setup';
