+++ packages/frontend/@n8n/i18n/src/__tests__/setup.ts
@@ -1,2 +1,1 @@
-// Avoid tests failing because of difference between local and GitHub actions timezone
-process.env.TZ = 'UTC';
+import '@n8n/vitest-config/frontend-setup';
