+++ packages/frontend/@n8n/stores/src/__tests__/setup.ts
@@ -1,7 +1,1 @@
-import '@testing-library/jest-dom/vitest';
-import { configure } from '@testing-library/vue';
-
-// Avoid tests failing because of difference between local and GitHub actions timezone
-process.env.TZ = 'UTC';
-
-configure({ testIdAttribute: 'data-test-id' });
+import '@n8n/vitest-config/frontend-setup';
